Update supports color to 5.3
