+++ types/supports-color/index.d.ts
@@ -1,15 +1,18 @@
-// Type definitions for supports-color 3.1.2
+// Type definitions for supports-color 5.3
 // Project: https://github.com/chalk/supports-color
-// Definitions by: Melvin Groenhoff <https://github.com/mgroenhoff>
+// Definitions by: Melvin Groenhoff <https://github.com/mgroenhoff>, Matt Traynham <https://github.com/mtraynham>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
-declare interface SupportsColor {
-    level: number;
-    hasBasic: boolean;
-    has256: boolean;
-    has16m: boolean;
+export namespace supportsColor {
+    interface Level {
+        level: number;
+        hasBasic: boolean;
+        has256: boolean;
+        has16m: boolean;
+    }
+
+    type SupportsColor = boolean & Level;
 }
 
-declare const supportsColor: boolean & SupportsColor;
-
-export = supportsColor;
+export const stdout: supportsColor.SupportsColor;
+export const stderr: supportsColor.SupportsColor;
