+++ types/supports-color/supports-color-tests.ts
@@ -1,17 +1,33 @@
-import supportsColor = require("supports-color");
+import { stdout, stderr } from "supports-color";
 
-if (supportsColor) {
-    // Terminal supports color
+if (stdout) {
+    // Terminal standard output supports color
 }
 
-if (supportsColor.hasBasic) {
-    // Terminal supports color
+if (stdout.hasBasic) {
+    // Terminal standard output supports color
 }
 
-if (supportsColor.has256) {
-    // Terminal supports 256 colors
+if (stdout.has256) {
+    // Terminal standard output supports 256 colors
 }
 
-if (supportsColor.has16m) {
-    // Terminal supports 16 million colors (truecolor)
+if (stdout.has16m) {
+    // Terminal standard output supports 16 million colors (truecolor)
+}
+
+if (stderr) {
+    // Terminal standard error supports color
+}
+
+if (stderr.hasBasic) {
+    // Terminal standard error supports color
+}
+
+if (stderr.has256) {
+    // Terminal standard error supports 256 colors
+}
+
+if (stderr.has16m) {
+    // Terminal standard error supports 16 million colors (truecolor)
 }
