+++ types/supports-color/v3/supports-color-tests.ts
@@ -1,0 +1,17 @@
+import supportsColor = require("supports-color");
+
+if (supportsColor) {
+    // Terminal supports color
+}
+
+if (supportsColor.hasBasic) {
+    // Terminal supports color
+}
+
+if (supportsColor.has256) {
+    // Terminal supports 256 colors
+}
+
+if (supportsColor.has16m) {
+    // Terminal supports 16 million colors (truecolor)
+}
