+++ types/supports-color/v3/tslint.json
